requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# NLP and embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector databases and search
faiss-cpu>=1.7.4
chromadb>=0.4.0
qdrant-client>=1.6.0

# Retrieval
rank-bm25>=0.2.2
nltk>=3.8.0
spacy>=3.6.0

# LLM integration
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-community>=0.0.10

# Evaluation
ragas>=0.1.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
colorama>=0.4.6
pyyaml>=6.0

# Development
pytest>=7.4.0
black>=23.7.0
